CMS-13138 [Backport 14.4] Add LastVisited observable model to BrowsePerspective
The browse perspective registers an observable LastVisited model and
registers itself as an observer. If the value changes, the change is
recorded in the BrowseState. When the BrowsePerspective renders, it
checks if last-visited has changed, and if so, calls #updateNavLocation.

The actual value change is handled by the BrowseService, which has a
listener for both a folder and a document change. The BrowseService get
a reference to the LastVisited observable model from the Navigator. To
prevent updates to the last-visited node being captured from navigating
a dialog, the Navigator checks if it is rendered inside a dialog and if
so passes a null reference instead of the last-visited model.

(cherry picked from commit 3434630eff74a0d6ddfa8eb509fd697c585386d1)