CMS-13101 Check for double items in put
Check for double items in put intead of when building the MenuLabels.
That whole stream does not add any value anymore, so that part has been
reverted to the "old style".

Add extra check for startsWith "info". The functionality only applies to
"info" items. The TranslationsPlugin for example depends on it that
multiple document variants add a ActionDescription ( that have the very
same id ). Possibly there are other cases as well. Best to be as
restrictive as possible.