CMS-12494 Add NoDoubleClick link & behavior
Most buttons and links in the CMS are single-click buttons. To prevent
users from either double-clicking or clicking the same button multiple
times will a request is in flight, we use a NoDoubleClick link or
behavior. For <input> and <button> elements, this is as simple as adding
the "disabled" attribute, for <a> we use a CSS workaround which disables
mouse events on the target element by adding "pointer-event: none;".