CMS-12551 Modify list cell based on retainability
Also get draft variant if it exists.
Modify cssClass and summary based on retainability.
Add and update translations for the state summary.
Add and update css classes for document states.

Do not modify the way the state summary is determined by the derived
data engine, because it's used by scxml and on many other places, that
would blow up the scope of this issue. Downside it that there are
translations of the "state summary" property that don't exist as a value
of the property.

Do not use the DocumentHandle or BranchHandle to get the draft variant,
because this piece of code has to be optimized for performance. The
DocumentHandle or BranchHandle. See comment on line 188.