CMS-12551 Always set icons
Always set empty icons, even if there is no published or draft variant.
If a document is created ( only a draft variant ), the list cell does
not have any icons. If the document is saved as draft, one of the icons
has to be replaced. However HippoIconStack#replaceFromSprite will throw
a NullPointerException, because there is no old icon.

That could not happen before, because the draft variant did not cause
the icons to change.