CMS-12551 Observe draft variant as well
Observe draft variant as well, because if the draft variant changes, the
tooltip and the icon of the ListCell has to change.

Could not add the handle as observable, because JcrNodeModel does not
recursively monitor changes.

Account for the case that there is only a draft ( use the state summary
from the draft document is that case ).

If the state is "changed", the document is "live" and there are
unpublished changes. If that state occurs in combination with a
retainable document, the state becomes "live-draft-changes", because the
fact that there are draft changes has preference. It guides the user to
the next action.