CMS-13158 [Backport 14.x] Make sure that actions on XPage is only allowed when editable
[Backport 14.x Experience Pages]

If the XPage is not editable, another user is most likely editing the
XPage. In this case, a backend action for modifying a container or
component in the unpublished variant should not succeed.

This is confirmed now in the integration tests

(cherry picked from commit d15d42324c2750c4178cbe9d6bda5ba8a93e0a38)