CMS-13415 Make transfer of owner possible
Move the invocation of "editDraft" after "saveDraft" from the server
side to the client side, so the client side can only "saveDraft" if the
right side panel closes or another perspective is selected.

The client side has to be able to only invoke keepDraft (not followed
by editDraft) when the user stops editing, so the document becomes
transferable.  Also the client side should open the document (editDraft
) immediately after saveDraft when the user keeps the draft and
continues editing, so that the hints reflect the latest state.  That way
correct server side validation, button enablement and user feedback is
assured.  This is completely anologous to the behaviour in the content
editor.  However the visual editor does not close the editor after
keepDraft, while the content editor does.

Starting state of the document:

/the-medusa-news:
  jcr:primaryType: hippo:handle
  /the-medusa-news[1]:
  /the-medusa-news[3]:
    hippostd:retainable: false
    hippostd:state: draft
    hippostd:transferable: false

hints:
obtainEditableInstance:true
commitEditableInstance:false
keepDraft:false
editDraft:false

User U1 starts editing the document:

/the-medusa-news:
  jcr:primaryType: hippo:handle
  /the-medusa-news[1]:
  /the-medusa-news[3]:
    hippostd:holder: U1
    hippostd:retainable: false
    hippostd:state: draft
    hippostd:transferable: false

hints after action:
obtainEditableInstance:false
commitEditableInstance:true
keepDraft:true
editDraft:false

User U1 saves the document as draft:

/the-medusa-news:
  jcr:primaryType: hippo:handle
  /the-medusa-news[1]:
  /the-medusa-news[3]:
    hippostd:holder: U1
    hippostd:retainable: true
    hippostd:state: draft
    hippostd:transferable: true

hints after action:
obtainEditableInstance:true
commitEditableInstance:false
keepDraft:false
editDraft:true

User U2 edits the draft:

/the-medusa-news:
  jcr:primaryType: hippo:handle
  /the-medusa-news[1]:
  /the-medusa-news[3]:
    hippostd:holder: U2
    hippostd:retainable: true
    hippostd:state: draft
    hippostd:transferable: false

hints after action:
obtainEditableInstance:false
commitEditableInstance:true
keepDraft:true
editDraft:false

Only if a document is transferable another user can take ownership of
the document.  A Document is only transferable when the document is not
edited and the document has been saved as draft.

Before this commit when following the sequence described in the issue,
the editor could not take ownership, because the document was not
transferable. (The editDraft hint is false and the document has an owner
other than the current user, the transferable property on the draft
variant is false ). The document was not transferable, because the editDraft
action was performed immediately after each saveDraft action ( server
side ). The editDraft action changes the state of the document from
transferable to non-transferable.  This would happen with any user
trying to take ownership of the document.