CMS-12551 Improve performance
Improve performance by only observerving one variant. That variant is
the draft variant if it exists otherwise the published or the
unpublished variant.

The variant is observed, because the tooltip and the icon have to change
if:
- the stateSummary changes
- the retainable property on the draft node changes

The derived data engine will update all variants if the state changes,
also the draft variant. So if the draft variant is observed if it
exists, only one observable is needed.