CMS-13373 [Backport 14.x] Add support for drag-drop groups
HST-Experience-Page-Components are not allowed to be dropped in
containers that are not marked with HST-Experience-Page-Component.

To implement this I've added a simple check to see if the source
container has the same drop-groups configured as the target container.
When starting a drag action, all container that are not in the same
drop-group are marked with a red overlay.