CMS-13424 Keep branchIdModel
The branchIdModel has to be kept when cancel edit ( discarding the
changes, disposeEditableInstance ), because custom and enterprise
plugins depend on the presence of the model to get e.g. a list of all
document branches.
This behaviour should be the save as for the other actions ( save, and
done ) in the EditingWorkflowPlugin.