CMS-13419 [Backport 14.3] By default, serve whitelisted resource regardless authenticated
When an SPA is rendered in the CM, this is typically done from a
different domain than the CMS runs on. The SPA in the CM however needs
to load cms angular channel manager resources : This is thus cross-origin.

However, cms angular channel manager resources are only allowed to be
loaded for logged in CMS users, see CMS-10733 : Most likely the secure
resources were introduces since the resources are loaded from a jar,
and we do not want for example .class files to be possible to request
for anonymous users.

Because of the Secure resource servlet, the SPA does these requests
with SameSite flag such that credentials cookies are send along with
the request (from SPA domain in browser a call to CMS domain). Then the
secure resource servlet returns the resources with

resp.setHeader("Access-Control-Allow-Origin", origin);
resp.setHeader("Access-Control-Allow-Credentials", "true");
resp.setHeader("Vary", "Origin");

Currently, for the CMS domain http session, we do not set any SameSite
value. Chrome will change the default SameSite behavior to 'Lax' when
not set in the future. From then on, the credentails will only be
included for top-level GET requests. We could fix it by setting
SameSite = None for our CMS http session, but that opens up
vulnerabilities.

This commit fixes the problem to by default not require an authenticated
cms user. Since the previously released SDK does include sending
credentials, we keep the CORS headers, otherwise the old SDK does not
work any more. Newer versions of the SDK can skip sending the
CMS credentials from the SPA

[1] https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Set-Cookie/SameSite

(cherry picked from commit 15f089272b267955900c7615fc70c19a4c06818f)