CMS-13450 Add requiresReload field
Experience pages can be viewed and modified in the channel manager.
You can, for example add a container item or move items in a
container. The node for such a page can be in version history
due to the projects feature. For such pages the back-end has to
perform additional steps before it can save the changes requested
by the front-end:
- first checkout the version to the unpublished variant under the handle
- then apply the changes to the hst:page node
- finally save the changes using the workflow session
As a result of these actions the uuids of child nodes of hst:page change.
This means that the HTML comments for the corresponding container items
are out of sync. To get back in sync the CM has to reload the page.

We will solve this by wrapping the ContainerItemRepresentation in an
ExtResponseRepresentation and add a boolean property to the representation
that indicates that the CM has to reload the page.

For consistency this is also done for the ContainerComponentResource although
that is not strictly necessary.

(cherry picked from commit 206e1db99786adba63f8603b9022f254d5cdbbf9)