CMS-13436 [Backport 14.x] Support sup prototypes and mixins to be added to a doc variant
[Backport 14.x Experience Pages]

when a new document is created, support in the 'arguments' two reserved
keys:

- extraMixins
- subPrototypeUUIDs

Both properties can be comma separated, where the first property is a
comma separated list of extra mixins and the second is a comma separated
list of uuids.

Since this feature is really targeted for intern usage, there are no
further constraints or check in the mixin type or uuids: If they are
invalid, just a repository exception is thrown. The practical use case
for this feature is the following:

When a new document is created and that document is an XPage, then after
the document prototype has been used to create the document, a second
step can add extra mixin types and extra sub-prototypes below the doc
variant. For experience pages, this would mean the mixin hst:page and
the hst:page node to be copied as subnode below the doc variant. Also
note that the nodename from the sub-prototype will be copied as is, no
replacement possible

(cherry picked from commit 5cd0d5b33ff04a02c2a3a86bfecda2f104da4d37)