CMS-13373 [Backport 14.x] Render 2 "edit-shared" buttons for filled containers
If a containers contains items, we render two buttons to switch between
editing "shared" and "local" containers: one button on the top of the
container and one on the bottom.

The button style is updated according to the design. Since we don't have
proper MD buttons in the iframe, we use the same styling as the CMS.