CMS-13373 [Backport 14.x] Do not render any html comments if user not in right role
[Backport 14.x Experience Pages]

When the user is not in the right role, (s)he won't be able to edit the
component whether it is locked or not. Hence, adding html comments is
pointless: (s)he doesn't even need to know whether some component is
locked by whom. As a result, the 'editable' flag is redundant: When
a component is not editable because of the user not being in the right
role, we just don't render html comment

(cherry picked from commit 5d31e9ce1ac91a1d2eb6af1365436f84a866e1a6)