CMS-13470 [Backport 14.x] Support getting the component item properties for frozen nodes
[Backport 14.x Experience Pages]

When an XPage gets rendered from version history, then when clicking on
the Hst component item in the CM, should show the parameters for the
frozen hst component item.

Therefore, when trying to find the unpublished variant in
CXFJaxrsHstConfigService take into account that the identifier from the
request can be from a frozen XPage node. Take this into account
when trying to find the XPage unpublished variant. Next to that,
when remapping the request, remap to jcr:frozenPrimaryType in case for
a frozen node

In PrivilegesAllowedInvokerPreprocessor for experience pages, we cannot
check the privileges any more against the hst component node sinec this
might be a frozen node (on which for example an author does not have role
hippo:author). Instead, check the privilege against the workspace handle
node.

In ParametersInfoAnnotationUtils, also support to load ParametersInfo
class if the component node is a frozen node

Lastly in XPageContainerItemComponentResourceTest validate that properties
from a versioned node can be read

(cherry picked from commit 2dc46f7e38a4d1fed0ef29ec748671fdc3be85bb)