CMS-13373 [Backport 14.x] Make sure that Page Meta Data is always added
[Backport 14.x Experience Pages]

After changes from

'Do not render any html comments if user not in right role'
commit e6f2b967de7daa7560fc5aefbbd8cf3bdd15daf6,

the result was that when compConfig.isInherited() = true,
no more html comment was rendered. However, the page meta
data for the top-level hst config should still be included.

However, this did not work for pages outside the hst:workspace
anymore: reason is that for example the home page typically is
like this:

+ myproject
    + hst:pages
        + homepage
    + hst:workspace
+ myproject-preview
    + hst:workspace

in preview, only the workspace is copied, hence the 'homepage'
config is marked as 'inherited', which is correct! But of course
still page meta data should be added.

This is now fixed with this commit making sure page meta data
is always added for the top hst response (root comp config)

(cherry picked from commit 060f9cd02e5ad9399fb1d11a3f7011d4bf2e1c30)