CMS-13373 [Backport 14.x] Make sure that Page Meta Data is always added
[Backport 14.x Experience Pages]

After changes from

'Do not render any html comments if user not in right role'
commit 7b5c4e1c2ffb1989cabd79bec29cfff159de4de7,

the result was that when compConfig.isInherited() = true,
no more html comment was rendered. However, the page meta
data for the top-level hst config should still be included.

However, this did not work for pages outside the hst:workspace
anymore: reason is that for example the home page typically is
like this:

+ myproject
    + hst:pages
        + homepage
    + hst:workspace
+ myproject-preview
    + hst:workspace

in preview, only the workspace is copied, hence the 'homepage'
config is marked as 'inherited', which is correct! But of course
still page meta data should be added.

This is now fixed with this commit making sure page meta data
is always added for the top hst response (root comp config)

(cherry picked from commit 6f6200865aeefb06689c40640194281f994a7a0a)