CMS-13450 Remove feedback service from container service
The ContainerService would show an error toast if the call to
PageStructureService.addComponentToContainer failed. But
PageStructureService.addComponentToContainer already shows an error
with a specific distinction between locked and generic errors. That
message was pushed away by the ContainerService and replaced
with a more generic message, which is a bug. Solved by removing
it.

(cherry picked from commit 392729b022cff20aee2226097b4c936f2db20d7f)