CMS-13450 Remove feedback service from container service
The ContainerService would show an error toast if the call to
PageStructureService.addComponentToContainer failed. But
PageStructureService.addComponentToContainer already shows an error
with a specific distinction between locked and generic errors. That
message was pushed away by the ContainerService and replaced
with a more generic message, which is a bug. Solved by removing
it.

(cherry picked from commit 4e1501152325bee01104ebd249845b37853512f9)