CMS-13450 Reload the iframe if the server requires it
Adding a new HST component to a container can result in an invalid
state on the client. The server will return "reloadRequired=true" in
this case, so we pass it from the PageStructureService to the
ContainerService and reload the iframe accordingly.

(cherry picked from commit ebcd29e739e5aa419f4e9ac90efda8fd127bb0ea)