CMS-13450 Reload the iframe if the server requires it
Adding a new HST component to a container can result in an invalid
state on the client. The server will return "reloadRequired=true" in
this case, so we pass it from the PageStructureService to the
ContainerService and reload the iframe accordingly.

(cherry picked from commit 6629a855af243d336754a8e5d80223245b98950a)