CMS-13450 Refetch the page instance after adding a component
The clientside state can become invalid when adding a new component to
a container. In this case, the iframe is reloaded and the "page"
reference becomes invalid, so we always grab the latest before we call
page.getComponentById.

(cherry picked from commit 04bba241e2220619b3e131b4ef92535588b3fd48)