CMS-13373 [Backport 14.x] For XPages, always check the handle for privileges
[Backport 14.x Experience Pages]

Before this fix, we'd check the component below an XPage whether it had
the right privilege (hippo:author) to find out whether the current user
is allowed to modify the XPage in the CM. However, this turns out to be
brittle, while we can just as well check the privilege on the containing
handle node, the actual document instead of the descendant XPage component

What went wrong is the following : Assume the following handle/doc
fixture

+ mydoc
  + mydoc (published)
     + hst:page
       + main
         + container
           + banner1
  + mydoc (unpublished)
     + hst:page
       + main
         + container
           + banner1
           + banner2

The unpublished has added a 'banner2' item which is not present below
the published version.

When the CMS user renders the request via the HST in the CM, this is
done with the 'preview user + (cms user - live/draft variants)'. As
such, when the HST Configuration is loaded for the unpublished, the
jcr session feteching the unpublished doc does not 'see'

mydoc/mydoc[2]

but just

mydoc/mydoc

This is because the published version is not readable. As a result, the
canonical stored location of the HST Component config in the XPage HST
page model for the unpublished is

mydoc/mydoc/hst:page/main/container/banner2

instead of the location

mydoc/mydoc[2]/hst:page/main/container/banner2

which is would be for the CMS User JCR Session without the removal of the
read access to the published and draft variants!

The isUserInRole for the CMS user against

'mydoc/mydoc/hst:page/main/container/banner2'

cannot be found since it does not exist for the CMS user! Only

'mydoc/mydoc[2]/hst:page/main/container/banner2'

exists!

To avoid this problem altogether since it is really hard to solve, just
check the privileges agains the handle instead!

(cherry picked from commit 4508df9908fe2595bf131e0bc42c172b58bd388a)