CMS-13486 Create navapp frontend-app
Currently the Navapp only needs one service to operate
namely the NavAppSettingsService. At the moment however
all CMS plugins and services are also loaded for the
nvapp which is a waste of (memory) server resources.
So to reduce the memory footprint of logged in users
only load the plugins that the navapp needs.

In other places we already check if the page being served
by Wicket is the navapp variant of the plugin page. So we
can make use of this to get the plugin service for the navapp.

We need a NavAppPlugin as component because every frontend
plugin application must be a Wicket component. For now this
component does not have it's own HTML, but I think it should
be possible to simplify it even more so that we don't need all
the tricks of excluding navapp resources from cms-app and vice
versa. I think that is best dealt with in a separate follow-up
issue.

(cherry picked from commit e804b44a80358fa75187eef41aae91c7b6cb0cf9)