CMS-13470 [Backport 14.x] Support modifying container and items from version history
[Backport 14.x Experience Pages]

When a CMS User invokes an action in the CM that targets a XPage which
is rendered from version history, make sure that first the XPage is
checked out from version history, and that consecutively, the interactions
are done against the *workspace* version of the XPage.

For this to work, the jcr:frozenUUID is checked from the container or
container items in version history and then the workspace version is
fetched via jcr:frozenUUID: this workspace version should always exist
since if needed, first

XPageUtils#checkoutCorrectBranch

if invoked which makes sure that the right branch is checked out first.

In case a version was checked out from version history, the response
has a representation that indicates requiresReload = true

There are integration tests covering most use cases. Only use case that
is not covered which might have to be covered later on as well is
interacting with a 'branch' instead of with 'master' : All version
history related tests are now such that 'master' is in version history,
and that the unpublished is for branch 'foo'.

(cherry picked from commit 8c8d7df4ddd4740fc515fb6057c10d1c96314416)