CMS-13473 [Backport 14.x] Share "edit-shared-containers" state with CM
When a page is reloaded, the angular app inside the iframe is restarted.
This means any state is reset, but this breaks the UX flow when editing
shared containers (which can trigger a page reload). So we have to share
and read the "isEditingSharedContainers" state with/from the CM by using
the communication service.

The flag "editing-shared-conatiners" is reset when a different page is
loaded.