CMS-13470 [Backport 14.x] Support XPage branches from version history being editted
[Backport 14.x Experience Pages]

The 'resty' endpoints for the HST need to be able to find out which
branch is selected on the CmsSessionContext to be able to perform
the action against the right branch. For this, the BranchSelectionService
which is used to find out the selected branch is needed : a
downstream project can inject a different BranchSelectionService, so
the only way for the HST to 'know' it is to be able to get hold of
the BranchSelectionService. It is added to the 'internal' package
since it is not meant as a public api.

Added integration tests which validate that both when the Master
branch or eg 'foo' branch are in version history, they can be both
edited in the CM

(cherry picked from commit f0503754570cb0122bd0fbbd433da974146e828a)