CMS-13416 Introduce ActionProvider
An action has a name, a category and an enabled flag. The
category represents the type of entity it belongs to (channel,
page or xpage for now). Implementations of ActionProvider must
return a set of actions. The ActionService will collect all
actions of all providers and returns them as a map keyed by
Category name. Then this map is converted into a map of
CategoryRepresentations that can be serialized into the desired
JSON format. Note that this means that we cannot have nested
items any more and that a category does not have an enabled
flag. Makes sense to me: a Category is only present if there is
at least one Action for that Category.

We plan to let the ComponentResource return states as well next
to the actions. In preparation I have added ActionsRepresentation
so that later on we can create a ComponentsRepresentation with
actions and states per relevant component. To make (integration)
testing easier I have also added the query parameter "unwrapped".
That way I can convert the JSON back into ActionsRepresentation
which hugely simplifies asserting the response.

(cherry picked from commit 7b4289e1939bf352d8d786b8f1cae0de9c06927a)