CMS-13433 Show the correct error message
When no layout is a selected on error message is shown:
"XPage Layout" is required.
Associate the label with the dropdown so wicket shows the
translation value instead of the key in the error message.

With special thanks to Arthur for pointing me to this
wicket behaviour.

(cherry picked from commit 2481a9e104ae6771a508e1e782420aa06cf72713)