CMS-13433 Show the correct error message
When no layout is a selected on error message is shown:
"XPage Layout" is required.
Associate the label with the dropdown so wicket shows the
translation value instead of the key in the error message.

With special thanks to Arthur for pointing me to this
wicket behaviour.

(cherry picked from commit ebdba707bca895fe5566d2d63de012f7dde3bc85)