CMS-13512 Check if proxy endpoint is live, use filter chain otherwise
Apart from refactoring the code and introducing the ProxyConfig class,
the main change is in the ProxyFilter starting at line 121.

By wrapping the #openConnection() call in a try-catch block together
with a #connect() we will either end up in the catch block (dev server
not running, chain the call to serve a static resource), or we can
safely continue and do the proxy magic.
Note that we need to copyRequestHeaders() before calling connect(),
otherwise we will run into an IllegalStateException.

I also noticed that the default jarPathPrefix was "META-INFO", but that
should be "META-INF" and start with a slash. It is not really important,
as there will always be a non-default jarPathPrefix for the ProxyFilter.

[Backport 14.x]