CMS-13450 [Backport 14.x] Refactor ComponentEditor to open by id
The ComponentEditor should be able to open by id and fetch an instance
of the componentElement, instead of the  objects currently used.
This will simplify the code as we no longer need to keep a reference to
the container since it is already in the ComponentElement.
It also simplifies (re)opening the editor, as only a string ID is
required.