CMS-13433 Save the xpage layout under the draft
The FolderworkflowPlugin gets the channelId from a ChannelIdProvider.
The only implementing interface is the HintsChannelIdProvider that gets
the channelId from the workflow hints. The workflow reads the channelId
from the folder node it is associated with. If that folder has the mixin
hippostd:xpagefolder and a hippostd:channelid property, the value of
that property is added to the hints.
The FolderWorkflowPlugin creates a XPageLayoutProvider that provides the
layouts used in the dialog for adding a new document. For now that's
just a plain jcr provider that reads the XPageLayouts from a json string
on the hst:channelinfo node.
The interfaces ChannelIdProvider and XPageLayoutProvider make the code
extendible and changeable.
The dialog adds a XPageLayoutContainer that has a LoadableDetachable
model for the xpage layouts, so that the xpage layouts are only then
fetched when the actual dropdown is added and won't be fetched again.
If there are no layouts, the dropdown is hidden.
When the user presses Ok, the WorkflowPlugin calls the FolderWorkflow.
The selected XPageLayout is added to the arguments parameter of the
workflow call.
When there is a valid subprototype UUID, the workflow will add the
hst:xpage mixin to the newly created draft and copy the sub prototype.
The NoSuchItemException is rethrown to give more sensible.

The jar packaging is removed from the pom.xml, because maven promotes
convention over configuration.

(cherry picked from commit 6765e5d1ca4c8680b8b1c95e91583adf642e676d)