CMS-13416 Disable page category for xpage request
A request cannot be for a page and an xpage at the
same time. Note that the CXFJaxrsHstConfigService
modified the path by prefixing it with experiencepage
if the request is for an experience page. However,
the prefix should not be added for requests for the
ComponentResource.

(cherry picked from commit 15e208f033fe36201b5c67283f1e9812bab10259)