CMS-13416 Disable page category for xpage request
A request cannot be for a page and an xpage at the
same time. Note that the CXFJaxrsHstConfigService
modified the path by prefixing it with experiencepage
if the request is for an experience page. However,
the prefix should not be added for requests for the
ComponentResource.

(cherry picked from commit 6b00cf8d893f63bc9b8887c391565cea58a83f42)