CMS-13416 Disable page category for xpage request
A request cannot be for a page and an xpage at the
same time. Note that the CXFJaxrsHstConfigService
modified the path by prefixing it with experiencepage
if the request is for an experience page. However,
the prefix should not be added for requests for the
ComponentResource.

(cherry picked from commit 44bc105badd93d1f8ce1cd3d55464d9b90b6e211)