CMS-13416 Disable page category for xpage request
A request cannot be for a page and an xpage at the
same time. Note that the CXFJaxrsHstConfigService
modified the path by prefixing it with experiencepage
if the request is for an experience page. However,
the prefix should not be added for requests for the
ComponentResource.

(cherry picked from commit bb3aadb76ed9c94aa3f9f1e318db0eeb16dca005)