CMS-13523 [backport 14.x] bump to new hippo jackrabbit version for property UUID support
In Jackrabbit support has been added that when a String value is equal
to ${fn:new-uuid()} , then the value gets replaced with a UUID.

We need this to be able to create hst containers which have a qualifier
which is in sync between copies of the same container. For example we
can now add to a cnd nodetype this:

[hst:foo]
- hst:qualifier (string) = '${fn:new-uuid()}' autocreated

this results in a hst:qualifier property containing a UUID for every
newly created node, and when this node gets copied, the copy get the
same value. This is exactly what we need when we want to copy an
XPage container below an XPage document variant: Namely we want to
decouple the XPage document variant container from the layout

(cherry picked from commit 7aab108f862f7aa6e06c4a5d7bf98a89d374dfb5)