CMS-13523 [backport 14.x]  Support 'hst:xpages' storage section in hst:configuration
In the hst configuration we need to be able to store the XPage
configurations that are available for a channel. We need to be able to
store a configuration like this:

+ hst:configuration
  + myproject
    + hst:xpages
       + xpage1
          + main
             + container1
               - hst:qualifier = abcd
             + container2
                - hst:qualifier = dcba

This is support now via :

[hst:xpage] > hst:component, hippostd:skipdraft
- hst:xref (string)
- hst:label (string)

[hst:xpages] > nt:base, mix:referenceable
+ * (hst:xpage)

Since we already had 'hst:xpage' as mixin, this turned out to be confusing
hence 'hst:xpage' mixin has been renamed to 'hst:xpagemixin' which is the
mixin on a document variant to allow a child node of type hst:xpage

Since we want 'containers' in the HST for XPages to have a qualifier which
is stable across copies, the property definition

- hst:qualifier (string) = '${fn:new-uuid()}' autocreated

has been added to the hst:containercomponent nodetype

From the 'hst:containercomponent' nodetype the mandatory property 'hst:xtype'
has been removed: Namely when copying the 'hst:containercomponent' to an XPage
document variant, we do not want to copy the 'xtype' : the 'xtype' should be
dictated fully by the hst XPage configuration and not be copied to document
variant instances. If the property is missing, a container should default to
xtype = HST.vBox (as the most sensible default)

(cherry picked from commit 0fa51447d79e8cd618fdbeeb17e4dc40504e5d6e)