CMS-13523 [backport 14.x]  Support loading 'hst:xpages' section in hst:configuration
Support loading hst:xpages into the HST Configuration model.

Xpages are loaded similarly to hst:pages/hst:abstractpages/hst:components,
they are only of type hst:xpage as 'root hst component'.

Just like normal hst components, they can reference another page. For
example a typical XPage might look as follows:

+hst:xpages:
  +xpage1 [hst:xpage]
    - hst:referencecomponent: hst:abstractpages/basepage
    - hst:label: XPage 1
        + main [hst:component]
          + container1[hst:containercomponent]
            - hst:xtype: HST.vBox
            - hst:qualifier: aaaa
          + container2 hst:containercomponent]
            - hst:xtype: HST.vBox
            - hst:qualifier: bbbb

Note the 'hst:qualifier' is a uuid which gets created automatically
because the cnd contains

- hst:qualifier (string) = '${fn:new-uuid()}' autocreated

When such a XPage is *USED* in a document variant, the document variant
will get stored something like below:

+ mydoc (handle)
  + mydoc (variant)
    + hst:xpage [hst:xpage]
      - hst:xref : xpage1
      + aaaa [hst:containercomponent]
      + bbbb [hst:containercomponent]

as can be seen above, the 'hst:xpage' copy to the document variant does
NOt contain the 'hst:referencecomponent' but instead contains hst:xref
which contains the name of the XPage : XPage has a unique name within a
single channel

The containers 'container1' and 'container2' will be copied (so any present
container items will be copied to the document variant), but the stored name
will be of the hst:qualifier: This way, the XPage layout in the HST
configuration can be changed *after* an XPage has been used in a document variant,
but the layout changes can then still be reflected in the rendered XPage
without having to modify the XPage below a document variant: the layout and
container items have been decoupled. Also the 'hst:xtype' *WON'T* be copied to
the document variant : it is a layout artifact

(cherry picked from commit 6afcfb43a7fb060f994bf3ae82b18423b8eb7bad)