CMS-13523 [backport 14.x]  Support rendering of XPage Document for new XPage Layout setup
The new structure to support is as follows:

+ hst:configuration
  + myproject
     + hst:xpages
        + xpage1
            + main
              + container1
                - hst:qualifier = abcd
              + container2
                - hst:qualifier = dcba

+ mydoc
     + mydoc (preview)
         + hst:xpage
             - hst:pageref = xpage1
             +abcd
                  + banner1
                  + banner2
             +dcba
                  + banner3
                  + banner4

In the code and comments I frequently refer to XPage Layout and XPage
Document. The XPage Layout refers to the hst config xpage below hst:xpages
and the XPage Document refers to an XPage below a document variant!

This means that when rendering an XPage Document, we find the hst:xpage
below the document variant mydoc/mydoc. In there we have the property

- hst:pageref = xpage1

The HST then needs to do the following steps

1) From the HST Model, find xpage1 Page Layout
2) Clone the xpage1 from the HST Model in a detached xpage1 configuration:
Once detached for a specific request, we can start modifying it
3) Traverse through the 'xpage1' cloned Layout, and for every container
in below the XPage Layout, take the hst:qualifier, and find the container
in the XPage Document with the nodename equal to the hst:qualifier : then
replace in the xpage1' cloned Layout the container ID, stored location
and some other properties with the one from the XPage document and add
the component items from the XPage document. Note that the container
properties like xtype are not replaced and are kept from the Layout: this
way the XPage Layout can later be modified impacting all XPage documents
already created directly

Lastly a lot of code is for covering corner cases, like for example when
a new container gets added to an existing XPage Layout: what should then
happen for existing XPage Documents for that layout? Choice: the new
container is kept without possible container items in the XPage layout:
this way, the container becomes visible in the Channel Mgr. Since it must
be editable, it even gets shared = false, etc. When in the Channel Mgr
then an item is added to it, a NEW container should be added to the
XPage document

Vice versa, when an XPage layout gets a container removed, then if there
are existing XPage Documents that refer to the hst:qualifier of the
removed container, then those XPage document containers are ignored : we
can't do else since we would not know where to place them

In that case, a info log message will be done with context about which
container is skipped

(cherry picked from commit 85bc3e326c6c6019308f1ff08b69d5450cbeeb72)