CMS-13601 HTML escape data if used in elements
The front-end code dynamically creates HTML. It uses data
that is provided by the back-end. In general we should not
rely on the back-end to provide data in HTML escaped format.
So to prevent XSS vulnerabilities all data that is added as
content to an element must be HTML escaped.

(cherry picked from commit d5c1260d032558c16c97231b40d225f394656dc8)