CMS-13601 HTML escape data if used in elements
The front-end code dynamically creates HTML. It uses data
that is provided by the back-end. In general we should not
rely on the back-end to provide data in HTML escaped format.
So to prevent XSS vulnerabilities all data that is added as
content to an element must be HTML escaped.

(cherry picked from commit c32a1d8819213e4a83836d83e8be6e5e8da1632c)